Prevent requesting arbitrary file paths via socket
In certain situations, specially crafted HTTP GET requests on the
livereload socket (default: 35729) may cause any user readable file to
be sent over that socket.

Temporarily in this patch, requests for readable files now result in 403
HTTP error responses. The exception is of course the file
'./livereload.js'.

Security vulnerability example:

Accessing the socket on localhost:35729 and requesting:

  ./../../etc/passwd (note the single leading dot)

to be expanded to "../../../etc/passwd", which may effectively serve the
contents of /etc/passwd.