[tiny] README - rephrase warning for clarity
On first glance it looks like this is still vulnerable — or so I thought. I think this slight rephrase makes it more clear on first glance. Given this is a development tool, I think catering to "at first glance" makes sense, since in development we're thinking "just go go go!" (Vs in production where we make more measured choices; for me that means reading a README more slowly for production)